add room lesson cards to room dashboard
For now lessons are shown in room dashboard below tasks.
Order is as they are given by the backend.
Implementation also gives room data to lesson card
as the href consists of the roomId and the topic id.
default color is given by the room or within the lesson title card.
If no user role is defined the room dashboard will refer
to the missing content in general instead of specific tasks missing.
Action button for teacher will be shown if lesson is hidden.
Added more info to language file for lesson card.