BC-5642 - internal links in LinkElements on boards (#2907)
The collection of meta tags for internal pages of the project can not be done via scraping. Therefore a set of rules for types of urls and implementations to gather the needed data was implementedv (in the backend).
On the frontendside the actual link-title is composed using entity-type-information based on the received information.