BC-7313 - handle focus for create column action (#3276)
When several users with edit rights work on the same board and one creates a new column, this column should be in edit mode and the title be focused for this user but not for the other users. They should only see the new column appearing.

As a side change: we activated sockets on column boards for development-environments.