BC-7445 - make text inputs on the board smooth (#3303)
When entering texts into text-fields for e.g. alt-Texts or descriptions in elements, sometimes some letters get lost. Reason for this is updates from the server coming in - that were triggered by the user (him|her)self.
Solution: While the user is editing a textfield no update from the server is interpreted as more up-to-date than the current state of the input-field the user has... and by that gets ignored.

---------

Co-authored-by: Murat Merdoglu <64781656+muratmerdoglu-dp@users.noreply.github.com>