BC-8873 updated canvas dependency (#3529)
due to jest-jsdom and jsdom refer to canvas in version 2, it results in a conflict, but good news: those dependencies are optional, only kjua relies on it.
And kjua dosen't seem to care what version of canvas it gets.
thus I added the overwrite part

this way, we can get pre-build binaries thus easing development